Use service
lift card catalog to container so do not have to deal with race conditions of mutating a global

fix

use service!

Fix detail panel
