+++ packages/host/app/resources/module-contents.ts
@@ -1,28 +1,16 @@
-import { tracked } from '@glimmer/tracking';
-
-import { restartableTask } from 'ember-concurrency';
+import { service } from '@ember/service';
 
 import { Resource } from 'ember-resources';
 
-import { getAncestor, getField, isBaseDef } from '@cardstack/runtime-common';
-
 import {
-  ModuleSyntax,
   type PossibleCardOrFieldClass,
   type BaseDeclaration,
-  type Declaration,
-  isInternalReference,
-  isPossibleCardOrFieldClass,
 } from '@cardstack/runtime-common/module-syntax';
 
-import {
-  getCardType,
-  type CardType,
-} from '@cardstack/host/resources/card-type';
+import { type CardType } from '@cardstack/host/resources/card-type';
 
 import { Ready as ReadyFile } from '@cardstack/host/resources/file';
-
-import { importResource } from '@cardstack/host/resources/import';
+import DeclarationsService from '@cardstack/host/services/declarations-service';
 
 import { type BaseDef } from 'https://cardstack.com/base/card-api';
 
@@ -54,83 +42,17 @@ interface Args {
 }
 
 export class ModuleContentsResource extends Resource<Args> {
-  @tracked private _declarations: ModuleDeclaration[] = [];
+  @service declare declarationsService: DeclarationsService;
 
   get isLoading() {
-    return this.load.isRunning;
-  }
-
-  get declarations() {
-    return this._declarations;
+    return this.declarationsService.load.isRunning;
   }
 
   modify(_positional: never[], named: Args['named']) {
     let { executableFile } = named;
-    this.load.perform(executableFile);
+    console.log('loading module contents');
+    this.declarationsService.load.perform(executableFile);
   }
-
-  private load = restartableTask(async (executableFile: ReadyFile) => {
-    //==loading module
-    let moduleResource = importResource(this, () => executableFile.url);
-    await moduleResource.loaded;
-    let exportedCardsOrFields = Object.values(
-      moduleResource?.module || {},
-    ).filter(isBaseDef);
-
-    //==building declaration structure
-    // This loop
-    // - adds card type (not necessarily loaded)
-    // - includes card/field, either
-    //   - an exported card/field
-    //   - a card/field that was local but related to another card/field which was exported, e.g. inherited OR a field of the exported card/field
-    let moduleSyntax = new ModuleSyntax(executableFile.content);
-    let localCardsOrFields = collectLocalCardsOrFields(
-      moduleSyntax,
-      exportedCardsOrFields,
-    );
-
-    this._declarations = moduleSyntax.declarations.reduce(
-      (acc: ModuleDeclaration[], value: Declaration) => {
-        if (isPossibleCardOrFieldClass(value)) {
-          const cardOrField = exportedCardsOrFields.find(
-            (c) => c.name === value.localName,
-          );
-          if (cardOrField) {
-            return [
-              ...acc,
-              {
-                ...value,
-                cardOrField,
-                cardType: getCardType(
-                  this,
-                  () => cardOrField as typeof BaseDef,
-                ),
-              } as CardOrField & Partial<PossibleCardOrFieldClass>,
-            ];
-          } else {
-            if (localCardsOrFields.has(value)) {
-              let cardOrField = localCardsOrFields.get(value) as typeof BaseDef;
-              return [
-                ...acc,
-                {
-                  ...value,
-                  cardOrField,
-                  cardType: getCardType(this, () => cardOrField),
-                } as CardOrField & Partial<PossibleCardOrFieldClass>,
-              ];
-            }
-          }
-        }
-        if (value.exportedAs !== undefined) {
-          // some classes that look like cards may still be included,
-          // we should only non-card or fields which are exported
-          return [...acc, { ...value } as BaseDeclaration];
-        }
-        return acc;
-      },
-      [],
-    );
-  });
 }
 
 export function moduleContentsResource(
@@ -140,78 +62,4 @@ export function moduleContentsResource(
   return ModuleContentsResource.from(parent, () => ({
     named: args(),
   })) as unknown as ModuleContentsResource;
-}
-
-function collectLocalCardsOrFields(
-  moduleSyntax: ModuleSyntax,
-  exportedCardsOrFields: (typeof BaseDef)[],
-): Map<PossibleCardOrFieldClass, typeof BaseDef> {
-  const localCardsOrFields: Map<PossibleCardOrFieldClass, typeof BaseDef> =
-    new Map();
-  let possibleCardsOrFields = moduleSyntax.possibleCardsOrFields;
-
-  for (const value of moduleSyntax.declarations) {
-    const cardOrField = exportedCardsOrFields.find(
-      (c) => c.name === value.localName,
-    );
-
-    if (cardOrField !== undefined) {
-      findLocalAncestor(
-        value,
-        cardOrField,
-        possibleCardsOrFields,
-        localCardsOrFields,
-      );
-      findLocalField(
-        value,
-        cardOrField,
-        possibleCardsOrFields,
-        localCardsOrFields,
-      );
-    }
-  }
-
-  return localCardsOrFields;
-}
-
-function findLocalAncestor(
-  value: ModuleDeclaration,
-  cardOrField: typeof BaseDef,
-  possibleCardsOrFields: PossibleCardOrFieldClass[],
-  localCardsOrFields: Map<PossibleCardOrFieldClass, typeof BaseDef>,
-) {
-  if (isPossibleCardOrFieldClass(value) && isInternalReference(value.super)) {
-    const indexOfParent = value.super.classIndex;
-    if (indexOfParent === undefined) return;
-    const parentCardOrFieldClass = possibleCardsOrFields[indexOfParent];
-    const parentCardOrField = getAncestor(cardOrField);
-
-    if (parentCardOrField == undefined) return;
-    localCardsOrFields.set(parentCardOrFieldClass, parentCardOrField);
-  }
-}
-
-function findLocalField(
-  value: ModuleDeclaration,
-  cardOrField: typeof BaseDef,
-  possibleCardsOrFields: PossibleCardOrFieldClass[],
-  localCardsOrFields: Map<PossibleCardOrFieldClass, typeof BaseDef>,
-) {
-  if (isPossibleCardOrFieldClass(value)) {
-    if (value.possibleFields) {
-      for (const [fieldName, v] of value.possibleFields) {
-        if (isInternalReference(v.card)) {
-          const indexOfParentField = v.card.classIndex;
-          if (indexOfParentField === undefined) return;
-          const parentFieldClass = possibleCardsOrFields[indexOfParentField];
-          const localName = parentFieldClass.localName;
-
-          if (localName === undefined) return;
-          const field = getField(cardOrField, fieldName);
-          if (field === undefined || field.card === undefined) return;
-          localCardsOrFields.set(parentFieldClass, field.card);
-        }
-      }
-    }
-  }
 }
